Commit message: Add RegisterView API endpoint with error handling
Explanation: In this commit, I added a new API endpoint called RegisterView that accepts HTTP POST requests. The endpoint takes in user data and uses the RegisterSerializer to validate the data. If the data is valid, the endpoint will create a new user account. If the data is invalid, the endpoint will return a JSON response with an error message and a 400 Bad Request status code.

I also added error handling to the endpoint to handle any unexpected exceptions that may occur during the processing of the request. If an exception occurs, the endpoint will return a JSON response with an error message and a 400 Bad Request status code.