Commit message: Update RegisterView to create new user accounts
Explanation: In this commit, I updated the RegisterView API endpoint to create new user accounts if the user data is valid. I added a call to the serializer.save() method, which creates a new user account based on the data provided in the request. If the user account is created successfully, the endpoint returns a JSON response with a success message and a 201 Created status code.

I also updated the error handling in the endpoint to include a comma after the serializer.errors() method call, which was missing in the previous commit. This was causing a syntax error in the code. Finally, I fixed a typo in the success message by changing "BAD" to "OK".