Added BaseModel as an abstract model with common fields and methods for all models.
Imported the uuid module to generate unique identifiers.
Imported the User model from django.contrib.auth.models.
Created the BaseModel class, inheriting from models.Model.
Added the uid field as a primary key of type UUIDField with the default value set to a generated UUID.
Added the created_at field as a DateField with auto_now_add=True, which sets the field to the current date and time when the object is created.
Added the updated_at field as a DateField with auto_now_add=True, which sets the field to the current date and time whenever the object is saved.
Created the Blog model, inheriting from BaseModel:

Added the user field as a ForeignKey to the User model, with on_delete=models.CASCADE to delete related blogs when a user is deleted, and related_name="blogs" to allow reverse querying from the User model.
Added the title field as a CharField with a maximum length of 200 characters.
Added the blog_text field as a TextField.
This commit provides a base model with common fields and methods and implements the Blog model.