Defined a BlogView class that extends the APIView class from the rest_framework.views module.
Set the permission_classes attribute to [IsAuthenticated], which means that only authenticated users will have permission to access this view.

Set the authentication_classes attribute to [JWTAuthentication], which specifies that JSON Web Token authentication will be used for this view.

Implemented a post method within the BlogView class. This method handles HTTP POST requests sent to the view.

Inside the post method, the request data is retrieved using request.data. Then, a BlogSerializer instance is created, passing the data to it.

If an exception occurs during this process, the exception is caught and printed for debugging purposes.