[Feature/Issue Number] [Summary of the change] [Actions taken to implement the change] [Additional details]
Commit Message:
"Update response data in 'get' method to use serialized data

This commit modifies the 'get' method to use the serialized data from the serializer instead of serializer.errors for the response data. The previous implementation incorrectly used serializer.errors, which returns any validation errors encountered during serialization, rather than the actual serialized data.

To address this issue, the following actions were taken:

Updated the 'data' key in the response dictionary to use serializer.data, which provides the serialized representation of the fetched blogs.
Removed the incorrect usage of serializer.errors, which is intended for validation errors and not for providing the serialized data.
This change ensures that the response data returned by the 'get' method includes the serialized representation of the fetched blogs. The serialized data contains the relevant fields and relationships from the Blog model, allowing clients to consume and process the data accurately.