Refactor 'patch' method for updating a blog with error handling
The 'patch' method has been refactored to handle updating a blog with proper error handling. The method retrieves the requested blog based on the UID provided in the request data. If the blog doesn't exist, it returns an error response indicating an invalid UID. If the requesting user is not authorized to update the blog, it returns an error response with an appropriate message.

The blog data is serialized using the 'BlogSerializer', and if the serialization is successful, the updated blog is saved. Finally, a success response is returned with the updated blog data.

In case of a 'ValueError', an error response is returned with a message indicating an invalid data format. For other exceptions, an error response is returned with a generic error message.

Note: Consider using a logging library instead of printing exceptions to the console.

Commit message: Refactor 'patch' method for updating a blog with error handling