Refactor 'delete' method for deleting a blog with error handling
The 'delete' method has been refactored to handle the deletion of a blog with proper error handling. The method retrieves the requested blog based on the UID provided in the request data. If the blog doesn't exist, it returns an error response indicating an invalid UID. If the requesting user is not authorized to delete the blog, it returns an error response with an appropriate message.

After successfully deleting the blog, a success response is returned with a message indicating the successful deletion.

In case of any exceptions, an error response is returned with a generic error message.

Note: Consider using a logging library instead of printing exceptions to the console.

Commit message: Refactor 'delete' method for deleting a blog with error handling