Refactor blog retrieval logic to use queryset method in views
In this commit, I refactored the blog retrieval logic in the views module. Instead of using a generic Python list to store the blogs, I utilized the all() queryset method provided by Django's ORM to fetch all blog objects from the database. This change improves the efficiency and readability of the code.

By leveraging queryset methods, we can take advantage of Django's optimized database queries, resulting in faster and more efficient retrieval of blog objects. This commit sets the foundation for future enhancements and optimizations in the project.