Import Paginator from django.core.paginator module
In this commit, I added an import statement to the views module to include the Paginator class from the django.core.paginator module. This import is necessary to utilize the pagination functionality in the code.

By including the Paginator class, we gain access to powerful tools for dividing large datasets into manageable pages. This allows us to implement pagination and improve the performance and user experience of our application.