Retrieve all blogs in PublicBlog's get method
In this commit, I updated the get method in the PublicBlog class. The blogs variable now retrieves all the blog objects from the Blog model using the objects.all() method.

This change ensures that all available blogs are fetched when the API endpoint for public blogs is accessed. The subsequent code can then perform operations, such as filtering or pagination, on the retrieved blogs.