Adjust pagination to display 3 blog posts per page
In this commit, I updated the pagination in the get method of the PublicBlog class. The Paginator now takes the blogs queryset and sets the page size to 3, allowing 3 blog posts to be displayed per page.

By increasing the number of blog posts per page, we can provide users with a more comprehensive view of the content and reduce the number of pages they need to navigate through.