Update success response status code to HTTP_200_OK
In this commit, I updated the success response status code in the get method of the PublicBlog class to status.HTTP_200_OK. This status code indicates a successful request.

By returning status.HTTP_200_OK, we provide the appropriate response status code to indicate that the request was successful. This helps clients and API consumers interpret the response correctly.