Randomize blog retrieval order in PublicBlog's get method
In this commit, I added the .order_by('?') method to the Blog.objects.all() query in the get method of the PublicBlog class. This change randomizes the order of the retrieved blog objects.

By randomizing the order of the blogs, we can provide a varied and dynamic browsing experience for users. Each time the API endpoint is accessed, the blogs will be presented in a different order, adding an element of unpredictability.