Update pagination to display 5 blog posts per page
In this commit, I adjusted the pagination in the get method of the PublicBlog class. The Paginator now takes the blogs queryset and sets the page size to 5, allowing 5 blog posts to be displayed per page.

By increasing the number of blog posts per page, we can provide users with a larger set of content to browse through in a single page view, potentially reducing the need for frequent page navigation.